Extracted miniflare options logic so that it can be rerun on config changes
